Make sure that flags are handled properly
Before this commit, if Pidgin sends any flag among those at the
following url (except for the first 2), no notification will be
triggered. For example, if you receive a chat message that contains your
nick, the flag will be 0x0002 + 0x0020. This means that equality is not
the way to check for flags. This commit makes use of the bitwise-and
operator to check if a particular flag is included.