Only notify for personal chat messages
Before this commit, if you are connected to a chat (i.e., a conversation
with multiple persons) that receives a lot of messages, you will see a
lot of notifications. This is problematic because it makes it impossible
to focus on something else and it lowers the chances you see the
important notifications. This commit makes sure that only chat messages
with your nick mentioned are notified. Non-chat messages (i.e.,
conversations with 1 person) are not affected by this change.

I think this could be an option of the extension.

This commit is independent but can only be easily merged after accepting
pull request #5.